[#52] Refactor: 에러코드 도메인별 분리
