+++ src/main/java/app/tamingo/common/response/ErrorCode.java
@@ -9,11 +9,6 @@ import org.springframework.http.HttpStatus;
 @AllArgsConstructor
 public enum ErrorCode implements BaseCode {
 
-    // 자주 가는 장소
-    FAVORITE_PLACE_NOT_FOUND(HttpStatus.NOT_FOUND, "PLACE-001", "해당 장소를 찾을 수 없습니다."),
-    FAVORITE_PLACE_ALREADY_EXISTS(HttpStatus.BAD_REQUEST, "PLACE-002", "이미 등록된 장소명 또는 주소입니다."),
-
-    // 서버 에러
     INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "COMMON-001", "서버 오류가 발생했습니다."),
     INVALID_REQUEST(HttpStatus.BAD_REQUEST, "COMMON-002", "요청 파라미터가 올바르지 않습니다."),
     INVALID_DATE(HttpStatus.BAD_REQUEST, "DATE-001", "유효하지 않은 날짜입니다."),
