+++ src/main/java/app/tamingo/domain/auth/service/LogoutService.java
@@ -1,7 +1,7 @@
 package app.tamingo.domain.auth.service;
 
 import app.tamingo.common.exception.CustomException;
-import app.tamingo.common.response.error.AuthErrorCode;
+import app.tamingo.domain.auth.exception.AuthErrorCode;
 import app.tamingo.common.security.JwtTokenProvider;
 import app.tamingo.domain.auth.redis.RefreshToken;
 import app.tamingo.domain.auth.redis.RefreshTokenRepository;
