+++ src/main/java/app/tamingo/domain/onboarding/exception/OnboardingErrorCode.java
@@ -1,4 +1,4 @@
-package app.tamingo.common.response.error;
+package app.tamingo.domain.onboarding.exception;
 
 import app.tamingo.common.response.BaseCode;
 import lombok.AllArgsConstructor;
