+++ src/main/java/app/tamingo/domain/terms/exception/TermsErrorCode.java
@@ -1,4 +1,4 @@
-package app.tamingo.common.response.error;
+package app.tamingo.domain.terms.exception;
 
 import app.tamingo.common.response.BaseCode;
 import lombok.AllArgsConstructor;
