+++ src/main/java/app/tamingo/domain/user/exception/UserErrorCode.java
@@ -1,4 +1,4 @@
-package app.tamingo.common.response.error;
+package app.tamingo.domain.user.exception;
 
 import app.tamingo.common.response.BaseCode;
 import lombok.AllArgsConstructor;
